Connectors filterPermissions + refactor Settings/Managed DS page + Carve out PermissionTree
